Добавил декоратор для обработки ошибок в функциях, переименовал собственный модуль logger в log_module
