+++ lab5/main.py
@@ -1,13 +1,16 @@
 from BlackJackCLI import BlackJackCLI
-from logger import module_logger
+from log_module import module_logger, log_errors
+from loguru import logger
 
 
+@log_errors(logger)
 def main():
     main_logger = module_logger.bind(service="main")
     game = BlackJackCLI()
     main_logger.info("Start Game")
     game.play()
     main_logger.info("Finish game")
+
 
 
 if __name__ == "__main__":
